DAS-2347 - Move get_resampled_dimensions to grid.py to avoid circular dependencies when used.
