+++ harmony_regridding_service/regridding_service.py
@@ -13,6 +13,9 @@ from harmony_regridding_service.crs import (
     add_grid_mapping_metadata,
     write_grid_mappings,
 )
+from harmony_regridding_service.dimensions import (
+    get_resampled_dimension_pairs,
+)
 from harmony_regridding_service.file_io import (
     clone_variables,
     transfer_metadata,
@@ -21,7 +24,6 @@ from harmony_regridding_service.grid import compute_target_area
 from harmony_regridding_service.resample import (
     cache_resamplers,
     copy_resampled_dimension_variables,
-    get_resampled_dimension_pairs,
     resample_n_dimensional_variables,
     transfer_resampled_dimensions,
     unresampled_variables,
